+++ src/js/index.js
@@ -1,6 +1,6 @@
 import Cars from "./controller/carsName/Cars.js";
 import RacingCount from "./controller/racingCount/RacingCount.js";
-import { makeInitialView } from "./view/util/viewControl.js";
+import { makeInitialView } from "./view/viewControl.js";
 import { EVENT } from "./util/constants.js";
 import { logo, restartButton } from "./util/elements.js";
