+++ src/js/view/winnerViewControl.js
